test: move helpers from common.js to helpers.js
