+++ dev/upload.html
@@ -10,7 +10,7 @@
     <script type="module">
       import '@vaadin/upload';
       import '@vaadin/radio-group';
-      import { xhrCreator } from '@vaadin/upload/test/common.js';
+      import { xhrCreator } from '@vaadin/upload/test/helpers.js';
 
       const nextUpload = document.querySelector('#next-upload');
