+++ packages/context-menu/test/visual/material/context-menu.test.js
@@ -2,7 +2,7 @@ import { fixtureSync, nextRender } from '@vaadin/testing-helpers';
 import { visualDiff } from '@web/test-runner-visual-regression';
 import '../../../theme/material/vaadin-context-menu.js';
 import '../../not-animated-styles.js';
-import { openSubMenus } from '../../common.js';
+import { openSubMenus } from '../../helpers.js';
 
 describe('context-menu', () => {
   let element;
