+++ packages/custom-field/test/custom-field.test.js
@@ -2,7 +2,7 @@ import { expect } from '@esm-bundle/chai';
 import { fixtureSync, focusin, focusout } from '@vaadin/testing-helpers';
 import sinon from 'sinon';
 import '../src/vaadin-custom-field.js';
-import { dispatchChange } from './common.js';
+import { dispatchChange } from './helpers.js';
 
 describe('custom field', () => {
   let customField;
