+++ packages/custom-field/test/slot-wrapper.test.js
@@ -2,7 +2,7 @@ import { expect } from '@esm-bundle/chai';
 import { fixtureSync, nextFrame } from '@vaadin/testing-helpers';
 import '../src/vaadin-custom-field.js';
 import { html, PolymerElement } from '@polymer/polymer/polymer-element.js';
-import { dispatchChange } from './common.js';
+import { dispatchChange } from './helpers.js';
 
 class XField extends PolymerElement {
   static get template() {
