+++ packages/date-picker/test/basic.test.js
@@ -4,7 +4,7 @@ import { sendKeys } from '@web/test-runner-commands';
 import sinon from 'sinon';
 import '../src/vaadin-date-picker.js';
 import { parseDate } from '../src/vaadin-date-picker-helper.js';
-import { close, open, touchTap, waitForOverlayRender } from './common.js';
+import { close, open, touchTap, waitForOverlayRender } from './helpers.js';
 
 describe('basic features', () => {
   let datepicker, input;
