+++ packages/date-picker/test/theme-propagation.test.js
@@ -1,7 +1,7 @@
 import { expect } from '@esm-bundle/chai';
 import { fixtureSync, nextRender } from '@vaadin/testing-helpers';
 import '../src/vaadin-date-picker.js';
-import { open, waitForScrollToFinish } from './common.js';
+import { open, waitForScrollToFinish } from './helpers.js';
 
 describe('theme attribute', () => {
   let datepicker;
