+++ packages/date-time-picker/test/validation.test.js
@@ -3,7 +3,7 @@ import { aTimeout, fixtureSync, nextRender } from '@vaadin/testing-helpers';
 import { sendKeys } from '@web/test-runner-commands';
 import sinon from 'sinon';
 import '../vaadin-date-time-picker.js';
-import { waitForOverlayRender } from '@vaadin/date-picker/test/common.js';
+import { waitForOverlayRender } from '@vaadin/date-picker/test/helpers.js';
 
 class DateTimePicker2020Element extends customElements.get('vaadin-date-time-picker') {
   checkValidity() {
