+++ packages/field-highlighter/test/visual/common.js
@@ -1,5 +1,4 @@
 import { css, registerStyles } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { FieldHighlighter } from '../../src/vaadin-field-highlighter.js';
 
 /* Hide caret */
 registerStyles(
@@ -31,16 +30,3 @@ registerStyles(
     }
   `,
 );
-
-const users = [
-  { id: 'a', name: 'User', colorIndex: 0, fieldIndex: 2 },
-  { id: 'b', name: 'Moderator', colorIndex: 1, fieldIndex: 1 },
-  { id: 'c', name: 'Admin', colorIndex: 2, fieldIndex: 0 },
-];
-
-export const setUsers = (field) => {
-  FieldHighlighter.init(field);
-  FieldHighlighter.setUsers(field, users);
-  // Mimic focus to show highlight and badges
-  field.dispatchEvent(new CustomEvent('mouseenter'));
-};
