+++ packages/field-highlighter/test/visual/material/field-highlighter.test.js
@@ -10,7 +10,8 @@ import '@vaadin/radio-group/theme/material/vaadin-radio-group.js';
 import '@vaadin/text-area/theme/material/vaadin-text-area.js';
 import '@vaadin/text-field/theme/material/vaadin-text-field.js';
 import '../../../theme/material/vaadin-field-highlighter.js';
-import { setUsers } from '../common.js';
+import '../common.js';
+import { setUsers } from '../helpers.js';
 
 describe('field-highlighter', () => {
   let div, element;
