+++ packages/upload/test/a11y.test.js
@@ -2,7 +2,7 @@ import { expect } from '@esm-bundle/chai';
 import { fixtureSync } from '@vaadin/testing-helpers';
 import sinon from 'sinon';
 import '../vaadin-upload.js';
-import { createFile } from './common.js';
+import { createFile } from './helpers.js';
 
 const FAKE_FILE = createFile(100000, 'application/uknown');
