+++ packages/vertical-layout/test/vertical-layout.test.js
@@ -1,7 +1,7 @@
 import { expect } from '@esm-bundle/chai';
 import { fixtureSync } from '@vaadin/testing-helpers';
 import '../vaadin-vertical-layout.js';
-import { getComputedCSSPropertyValue } from './common.js';
+import { getComputedCSSPropertyValue } from './helpers.js';
 
 describe('vaadin-vertical-layout', () => {
   describe('basic features', () => {
